requirements.txt

einops==0.8.1
faiss-cpu==1.9.0.post1
fastapi==0.115.11
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
h11==0.14.0
hf-xet==1.1.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
langchain==0.3.16
langchain-community==0.3.16
langchain-core==0.3.32
langchain-google-genai==2.0.9
langchain-text-splitters==0.3.5
langsmith==0.3.13
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.38.0
networkx==3.5
nomic==3.5.3
numpy==2.2.3
pandas==2.2.3
pdfplumber==0.11.6
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.3
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==4.1.0
streamlit==1.45.1
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.4
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2

